+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## V0.1.2
+- Make configuration directory a constant
+- Change configuration directory from ~/.wp-sync-util to ~/.wp-site-sync
+- Add return code checking for ssh, scp and rsync commands
+- Add code to remove the remote backup file after it's copied locally
+
 ## V0.1.1
 - Added README.md file
 - Added CHANGELOg.md file